requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
crewai>=1.14.0
crewai-tools>=0.14.0
langchain-anthropic>=1.4.0
python-dotenv>=1.0.0
pydantic>=2.11.0
gunicorn==21.2.0
